allow to use custom translator & SDL2

+++ scripts/skps2010Scripts/makeMacOSX.sh
@@ -7,8 +7,12 @@ fi
 
 ./skps2010Build.sh
 
-python3 -m PyInstaller --onefile translator.py
-cp dist/translator .
+# allow to use custom translator
+if [ ! -e translator ]
+then
+   python3 -m PyInstaller --onefile translator.py
+   cp dist/translator .
+fi
 
 cd OneLife/build
 ./makeDistributionMacOSX v$1 _ /usr/local/opt/sdl12-compat/lib/libSDL-1.2.0.dylib
